Separated into several classes and moved to own package. Removed use of mocking library. Other misc fixes
