+++ metrics-proxy/pom.xml
@@ -149,11 +149,6 @@
       <artifactId>hamcrest-core</artifactId>
       <scope>test</scope>
     </dependency>
-      <dependency>
-        <groupId>org.mockito</groupId>
-        <artifactId>mockito-core</artifactId>
-        <scope>test</scope>
-      </dependency>
   </dependencies>
   <build>
     <plugins>
